% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnoPlexR-package.R
\docType{package}
\name{InnoPlexR-package}
\alias{InnoPlexR}
\alias{InnoPlexR-package}
\title{InnoPlexR: Build and Analyze Multiplex Networks}
\description{
Tools for constructing and analyzing multiplex networks. The package
provides a \code{Multiplex} S3 class built on top of igraph layers, along with
functions for computing matrix representations (supra-adjacency, supra-
Laplacian, supra-transition) and a suite of multilayer centrality measures.
}
\section{Core workflow}{

\enumerate{
\item Build a \code{Multiplex} object with \code{\link[=create_Multiplex]{create_Multiplex()}}.
\item Compute matrix representations with \code{\link[=compute_supra_adjacency_matrix]{compute_supra_adjacency_matrix()}},
\code{\link[=compute_supra_laplacian_matrix]{compute_supra_laplacian_matrix()}}, or \code{\link[=compute_supra_transition_matrix]{compute_supra_transition_matrix()}}.
\item Compute centralities: \code{\link[=GetMultiEigenvectorCentrality]{GetMultiEigenvectorCentrality()}},
\code{\link[=GetMultiPageRankCentrality]{GetMultiPageRankCentrality()}}, \code{\link[=GetMultiGeodesicBetweenness]{GetMultiGeodesicBetweenness()}},
\code{\link[=GetMultiRandomWalkBetweenness]{GetMultiRandomWalkBetweenness()}}.
\item Analyse structure: \code{\link[=GetMultiComponents]{GetMultiComponents()}}, \code{\link[=GetMultiLargestComponent]{GetMultiLargestComponent()}},
\code{\link[=GetLocalClustering]{GetLocalClustering()}}.
\item Export for external tools: \code{\link[=multiplex_to_infomap]{multiplex_to_infomap()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zabelmt/InnoPlexR}
  \item Report bugs at \url{https://github.com/zabelmt/InnoPlexR/issues}
}

}
\author{
\strong{Maintainer}: Marcus Zabel \email{m.t.zabel@students.nl}

}
\keyword{internal}
