% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplex_class.R
\name{is_Multiplex}
\alias{is_Multiplex}
\title{Test whether an object is a Multiplex}
\usage{
is_Multiplex(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
\code{TRUE} if \code{x} has class \code{"Multiplex"}, \code{FALSE} otherwise.
}
\description{
Test whether an object is a Multiplex
}
\examples{
g <- igraph::make_ring(5)
igraph::V(g)$name <- letters[1:5]
mx <- create_Multiplex(g, Layers_Name = "ring")
is_Multiplex(mx)   # TRUE
is_Multiplex(g)    # FALSE
}
