% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplex_class.R
\name{print.Multiplex}
\alias{print.Multiplex}
\title{Print a Multiplex object}
\usage{
\method{print}{Multiplex}(x, ...)
}
\arguments{
\item{x}{A \code{Multiplex} object.}

\item{...}{Currently unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a Multiplex object
}
