% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplex_class.R
\name{summary.Multiplex}
\alias{summary.Multiplex}
\title{Summarise a Multiplex object}
\usage{
\method{summary}{Multiplex}(object, ...)
}
\arguments{
\item{object}{A \code{Multiplex} object.}

\item{...}{Currently unused.}
}
\value{
A data frame (invisibly) with one row per layer.
}
\description{
Summarise a Multiplex object
}
